repro/requirements.txt
PyYAML==6.0
requests==2.28.2
rich==13.3.3
shortuuid==1.0.11
